+++ src/parking_lot.hh
@@ -16,20 +16,20 @@ private:
   Flow B { "B", "z" };
   Flow C { "C", "z" };
 
-  std::tuple<double, double, double> throughputs_shortcut( const double A_rate,
-							   const double B_rate,
-							   const double C_rate ) const;
+  std::tuple<float, float, float> throughputs_shortcut( const float A_rate,
+							   const float B_rate,
+							   const float C_rate ) const;
 
   unsigned int audit_count_ {};
 
 public:
-  std::tuple<double, double, double> throughputs( const double A_rate,
-						  const double B_rate,
-						  const double C_rate );
+  std::tuple<float, float, float> throughputs( const float A_rate,
+						  const float B_rate,
+						  const float C_rate );
 
-  std::tuple<double, double, double> throughputs_fast( const double A_rate,
-						       const double B_rate,
-						       const double C_rate,
+  std::tuple<float, float, float> throughputs_fast( const float A_rate,
+						       const float B_rate,
+						       const float C_rate,
 						       const unsigned int i );
 
   unsigned int audit_count() const { return audit_count_; }
